Macro multiversion::are_cpu_features_detected

macro_rules! are_cpu_features_detected {
    { $feature:tt $(,)? } => { ... };
    { $first:tt, $($features:tt),+ $(,)? } => { ... };
}

Detects CPU features.

When the std feature is enabled, this macro operates like the standard library detection macro for the current target (e.g. is_x86_feature_detected), but accepts multiple arguments.

When the std feature is not enabled, this macro detects if the feature is enabled during compilation, using the cfg attribute.